requirements.txt
streamlit
python-dotenv
pinecone
cohere
openai
beautifulsoup4
requests